drm/msm: use drm helper API for atomic check
Current msm atomic check tries atomic check first and
modeset later. That causes issue for crtc atomic check
because crtc mode is not updated so vdisplay/hdisplay
is "0" when atomic check happen before modeset. That leads
to atomic commit failure.

Originally, it was added for plane format checks before it
updates. That may not be correct with sde driver because
it handles all atomic checks independently of modeset changes.

Change-Id: Id8cbbcb2f520ac2956cf56cb959e120ae6f55855
Signed-off-by: Dhaval Patel <pdhaval@codeaurora.org>
Signed-off-by: MOVZX <movzx@yahoo.com>