binder: remove race condition in binder_dequeue_work
This race condition is a likely explanation for one
observed case where a stale work entry is left on
a proc->todo list causing a later assertion failure.
If two threads are dequeuing an entry at the same time
(eg, for release and for normal queue processing), one
of them could be acquiring the wrong lock if the
work was requeued on a different list by the other
thread. This could result in the wlist pointer being
NULL, but the work node still being on a list -- so
it will never dequeue.

Bug: 35302164
Change-Id: Ibeb150dfcfa79dfeda49596d021d3da41eb08774
Test: tested manually
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: Marco Zanin (B--B) <mrczn.bb@gmail.com>