binder: fix possible double kfree of ref->death
If a binder_thread is being released and has a BINDER_WORK_DEAD_BINDER
on its todo list, the corresponding death notification object is being
kfree'd as part of the cleanup, but since ref->death still points to
it, the zombie cleanup of the ref will also kfree it. Not sure if this
has happened, but it could.

Change-Id: Ic300922f4efbeb039657deaa8715c23bb6ca06ca
Test: tested manually
Bug: 32225111
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: Marco Zanin (B--B) <mrczn.bb@gmail.com>