binder: fix false BUG_ON
binder_enqueue_work asserts that the object is not
already queued by testing if work->wlist != NULL. The
assertion is outside the critical section, so it is
possible that the assertion can trip when the work
item is off the queue, but before work->wlist is
cleared (as was the case in b/36511858). Move the
assertion into the critical section so wlist check
is atomic with the actual list.

Bug: 36511858
Change-Id: I4d65e5abaa8a4bb0e3c122869ca8cca0991b83ed
Test: tested manually
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: Marco Zanin (B--B) <mrczn.bb@gmail.com>