ANDROID: binder: improve priority inheritance.
By setting the priority of a new incoming transaction
before waking up the thread that will handle it. Note
that this change doesn't affect the case where no
thread is available to handle work; we will fall
back to the old behavior in that case.

This change also delays restoring the priority in the
return path: it only restores the priority after the
original caller has been woken up.

Change-Id: I94034a28464ccd3ee16856d817497db1f4fb8e21
Signed-off-by: Martijn Coenen <maco@google.com>
Signed-off-by: Marco Zanin (B--B) <mrczn.bb@gmail.com>